\name{get.whites}
\alias{get.whites}
\title{
An internal function determining grayscale values}
\description{
Used internally in good.fibes to determine grayscale values along all possible line paths
}
\usage{
get.whites(imglist, lpl, depth, backstep = 0)
}
\arguments{
  \item{imglist}{
The current images being considered
}
  \item{lpl}{
generated by points.generator to find the position of all possible paths
}
  \item{depth}{
radius + 1 from good.fibes
}
  \item{backstep}{
How many images to consider "behind" the plane of the zero.image. 
}
}
\details{
\strong{for internal use in good.fibes only}
}
\value{
A matrix of grayscale values > cutoff
}
\keyword{internal}
