\name{GoodFibes-package}
\alias{GoodFibes-package}
\alias{GoodFibes}
\docType{package}
\title{
\packageTitle{GoodFibes}
}
\description{
\packageDescription{GoodFibes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GoodFibes}
\packageIndices{GoodFibes}
Functions in this package allow for the reconstruction of muscle fibers from diceCT image stacks. Fibers are detected using textural analysis, smoothed using splines and processed for path quality (grayscale variation). Muscle fibers can be plotted in 3D with or without the overall muscle outline, and the 3D muscle fibers can be exported as an STL file. Basic fiber architecture metrics can be calculated.}
\author{
\packageAuthor{GoodFibes}

Maintainer: \packageMaintainer{GoodFibes}
}
\references{
Arbour, J. In Prep. GoodFibes: an R package for the detection of muscle fibers from diceCT scans.
}
\keyword{ package }
