\name{pointsGenerator}
\alias{pointsGenerator}
\title{
Line points generator
}
\description{
An internal function for generating line coordinates for all possible paths of the good.fibes tracking algorithm.
}
\usage{
pointsGenerator(startx, starty, ucoords, radius = radius, backstep)
}
\arguments{
  \item{startx}{
seed point x coordinate
}
  \item{starty}{
seed point y coordinate
}
  \item{ucoords}{
unique end coordinates generated by hemisphere.points
}
  \item{radius}{
Number of images to consider forward from the zero.image
}
  \item{backstep}{
Should images in the seed point plane or behind it be considered. Each value (1, 2, 3, etc.) gives the number of images "behind" the seed point to consider in the spherical dome
}
}
\details{
For internal use in good.fibes only}
\value{
A list containing coorindates for each line ending in the end points determined by hemisphere.points
}
\keyword{ ~internal }
